Django-Middleware-0x03/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
drf-nested-routers==0.95.0
sqlparse==0.5.3
